Psm fdr revamp (#1481)
* fdr changes for psms that match to both a target and a decoy sequence

* fix bug, fix unit tests

+++ EngineLayer/FdrAnalysis/FdrInfo.cs
@@ -2,10 +2,10 @@
 {
     public class FdrInfo
     {
-        public int CumulativeTarget { get; set; }
-        public int CumulativeDecoy { get; set; }
-        public int CumulativeTargetNotch { get; set; }
-        public int CumulativeDecoyNotch { get; set; }
+        public double CumulativeTarget { get; set; }
+        public double CumulativeDecoy { get; set; }
+        public double CumulativeTargetNotch { get; set; }
+        public double CumulativeDecoyNotch { get; set; }
         public double QValue { get; set; }
         public double QValueNotch { get; set; }
         public bool CalculateEValue { get; set; }
